requirements.txt
# CEG4195 Lab 2 – Python dependencies
flask==2.3.3
python-dotenv==1.0.0
gunicorn==21.2.0

# Deep learning
torch==2.1.0
torchvision==0.16.0

# Image processing
Pillow==10.1.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
matplotlib==3.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
